+++ buy_sell_algorithm/optimization.py
@@ -1,12 +1,11 @@
 import numpy as np
 import cvxpy as cp
-import time
 import data.server_data as data
 import predictions.train as train
 
-def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffers, predictions_buffer, t, time_step, horizon):
+def maximize_profit_mpc(initial_storage_level, data_buffers, predictions_buffer, t, horizon):
 
-    #print(predictions_buffer)
+    MAX_STORAGE_CAPACITY = 50
     predicted_buy_prices = predictions_buffer['buy_price']
     predicted_sell_prices = predictions_buffer['sell_price']
     predicted_demand = predictions_buffer['demand']
@@ -20,7 +19,10 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
     current_buy_price = data_buffers['buy_price'][-1]
     current_sell_price = data_buffers['sell_price'][-1]
     energy_in = energy_in * 0.1 
-    #start, end, energy = get_deferables()
+
+    # Update predictions
+    predicted_buy_prices[t] = current_buy_price
+    predicted_sell_prices[t] = current_sell_price
 
     serve = data.server_data()
     serve.deferables()
@@ -46,13 +48,13 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
     neg_storage_transactions = cp.Variable(horizon, nonneg=True)
 
     # Objective function
-    profit = cp.sum(cp.multiply(neg_energy_transactions, predicted_sell_prices[t:t + horizon]) - cp.multiply(pos_energy_transactions, predicted_buy_prices[t:t + horizon]))
+    profit = cp.sum(cp.multiply(neg_energy_transactions, predicted_buy_prices[t:t + horizon]) - cp.multiply(pos_energy_transactions, predicted_sell_prices[t:t + horizon]))
 
     # Constraints
     constraints = [
         storage_level[0] == storage,  # storage level currently
         solar_energy[0] == energy_in,
-        demand[0] == energy_used,
+        demand[0] == energy_used,  # Set initial demand to energy used
         energy_transactions == pos_energy_transactions - neg_energy_transactions,
         storage_transactions == pos_storage_transactions - neg_storage_transactions,
         pos_energy_transactions + solar_energy - demand - neg_energy_transactions >= 0,
@@ -63,17 +65,16 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
     for i in range(horizon):
         constraints += [
             storage_level[i + 1] == storage_level[i] - storage_transactions[i],
-            storage_level[i + 1] <= max_storage_capacity,
+            storage_level[i + 1] <= MAX_STORAGE_CAPACITY,
             storage_level[i + 1] >= 0,
-            neg_energy_transactions[i] <= storage_level[i], # Can only sell if we have enough in the storage
-            pos_energy_transactions[i] <= max_storage_capacity - storage_level[i],  # Can only buy if we have enough storage capacity
+            neg_energy_transactions[i] <= storage_level[i],  # Can only sell if we have enough in the storage
+            pos_energy_transactions[i] <= MAX_STORAGE_CAPACITY - storage_level[i],  # Can only buy if we have enough storage capacity
         ]
     
     for i in range(1, horizon):  # Starting from 1 since solar_energy[0] is fixed to energy_in
         constraints += [
             solar_energy[i] == 0,  # Worst-case scenario
-            demand[i] == predicted_demand[i],  # Predicted demand
-            #storage_level[i] == storage_level[i-1] - storage_transactions[i-1],
+            demand[i] == predicted_demand[t + i],  # Set subsequent demands to predicted demand
         ]
 
     # Define problem
@@ -83,14 +84,9 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
     problem.solve(solver=cp.CBC)  # Using CBC mixed-integer solver (Coin-or branch and cut)
 
     if problem.status == cp.INFEASIBLE:
-        print(f" energy_transactions.value: {energy_transactions.value}")
-        print(f" storage_transactions.value: {storage_transactions.value}")
-        print(f" solar_energy: {solar_energy.value}")
-        print(f" demand: {demand.value}")
-        print(f" storage: {storage_level.value}")
-        print("Infeasible")
+        print("INFEASIBLE SOLUTION")
 
-        #use naive solution if solution is infeasible
+        # Use naive solution if solution is infeasible
         if energy_in >= energy_used:
             energy_in -= energy_used
             energy_used = 0
@@ -107,12 +103,21 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
                 storage = 0
         
         if energy_in > 0:
-            if storage + energy_in <= max_storage_capacity:
+            if storage + energy_in <= MAX_STORAGE_CAPACITY:
                 storage += energy_in
             else:
-                excess_energy = storage + energy_in - max_storage_capacity
+                excess_energy = storage + energy_in - MAX_STORAGE_CAPACITY
                 total_profit += excess_energy * current_sell_price
-                storage = max_storage_capacity
+                storage = MAX_STORAGE_CAPACITY
+
+        print(f" energy_transactions.value: {energy_transactions.value}")
+        print(f" storage_transactions.value: {storage_transactions.value}")
+        print(f" solar_energy: {solar_energy.value}")
+        print(f" demand: {demand.value}")
+        print(f" storage: {storage_level.value}")
+        print(f" Predicted Demand: {predicted_demand[t:t + horizon]}")
+        print(f" Predicted Buy Prices: {predicted_buy_prices[t:t + horizon]}")
+        print(f" Predicted Sell Prices: {predicted_sell_prices[t:t + horizon]}")
         
     elif problem.status == cp.OPTIMAL or problem.status == cp.FEASIBLE:
         optimal_energy_transaction = energy_transactions.value[0]   
@@ -126,7 +131,7 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
 
         # Ensure storage is within bounds after transactions
         storage -= optimal_storage_transaction
-        storage = min(max(storage, 0), max_storage_capacity)
+        storage = min(max(storage, 0), MAX_STORAGE_CAPACITY)
 
         print(f"  Energy Transaction: {optimal_energy_transaction} kWh")
         print(f"  Storage Transaction: {optimal_storage_transaction} kWh")
@@ -140,6 +145,7 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
         print(f" solar_energy: {solar_energy.value}")
         print(f" demand: {demand.value}")
         print(f" storage: {storage_level.value}")
+        print(f" Predicted Demand: {predicted_demand[t:t + horizon]}")
         print(f" Predicted Buy Prices: {predicted_buy_prices[t:t + horizon]}")
         print(f" Predicted Sell Prices: {predicted_sell_prices[t:t + horizon]}")
 
@@ -147,9 +153,4 @@ def maximize_profit_mpc(initial_storage_level, max_storage_capacity, data_buffer
         print(f"   Optimization failed")
         print(problem.status)
 
-    return total_profit
-
-
-
-if __name__ == '__main__':
-    print("Running optimization")
+    return total_profit, storage
